refactor(pool-page): extract OLS components to _components/
Move three OLS components out of page.tsx into _components/:

- ols-status-panel.tsx — OlsStatusPanel (~170 lines, exported)
- ols-liquidity-table.tsx — OlsLiquidityTable (~75 lines, exported)
- ols-liquidity-events.tsx — OlsLiquidityEvents (~150 lines, internal)

page.tsx imports all three and re-exports OlsStatusPanel +
OlsLiquidityTable so __tests__/ols.test.ts (which imports them from
"../page") keeps working without edits — pinned by exports.test.ts.

Behavior-preserving — verified via:
- pnpm test (1366/1366 green)
- pnpm typecheck (clean)
- chrome-devtools snapshot of /pool/<id>?tab=ols vs baseline:
  zero structural changes, console clean

page.tsx: 2746 → 2348 lines (-398, -15%).
