+++ ui-dashboard/src/app/pool/[poolId]/page.tsx
@@ -143,12 +143,17 @@ import {
   REWARD_TOOLTIP,
 } from "./_lib/tooltips";
 import type { OracleSortCol } from "./_lib/types";
+import { OlsLiquidityEvents } from "./_components/ols-liquidity-events";
+import { OlsLiquidityTable } from "./_components/ols-liquidity-table";
+import { OlsStatusPanel } from "./_components/ols-status-panel";
 
-// Re-export public helpers — `__tests__/ols.test.ts` and `page.test.tsx`
+// Re-export public symbols — `__tests__/ols.test.ts` and `page.test.tsx`
 // import these directly from "../page". Keep the import paths stable.
 export {
   decodePoolId,
   getDebtTokenSideLabel,
+  OlsLiquidityTable,
+  OlsStatusPanel,
   parseTabLimit,
   selectActiveOlsPool,
 };
@@ -2340,407 +2345,5 @@ function OlsTab({
         onSearchChange={onSearchChange}
       />
     </div>
-  );
-}
-
-export function OlsStatusPanel({
-  olsPool,
-  pool,
-  network,
-}: {
-  olsPool: OlsPool | null;
-  pool: Pool | null;
-  network: ReturnType<typeof useNetwork>["network"];
-}) {
-  if (!olsPool) {
-    return (
-      <div className="rounded-lg border border-slate-800 bg-slate-900/60 p-5">
-        <p className="text-slate-400 text-sm">
-          This pool is not registered with the Open Liquidity Strategy.
-        </p>
-      </div>
-    );
-  }
-
-  const nowSeconds = Math.floor(Date.now() / 1000);
-  const lastRebalance = Number(olsPool.lastRebalance);
-  const cooldown = Number(olsPool.rebalanceCooldown);
-  const elapsed = lastRebalance > 0 ? nowSeconds - lastRebalance : null;
-
-  const cooldownReady =
-    lastRebalance > 0 && elapsed !== null && elapsed >= cooldown;
-  const cooldownActive =
-    lastRebalance > 0 && elapsed !== null && elapsed < cooldown;
-  const cooldownPct =
-    cooldownActive && cooldown > 0
-      ? Math.max(0, Math.min((elapsed! / cooldown) * 100, 100))
-      : 0;
-  const cooldownRemaining = cooldownActive
-    ? Math.max(0, cooldown - elapsed!)
-    : 0;
-  const cooldownH = Math.floor(cooldownRemaining / 3600);
-  const cooldownM = Math.floor((cooldownRemaining % 3600) / 60);
-
-  const debtTokenSym = tokenSymbol(network, olsPool.debtToken || null);
-  const debtTokenSide = getDebtTokenSideLabel(pool, olsPool.debtToken);
-
-  return (
-    <div className="rounded-lg border border-slate-800 bg-slate-900/60 p-5 space-y-5">
-      <div className="flex flex-wrap items-center gap-3">
-        <h3 className="text-base font-semibold text-white">
-          Open Liquidity Strategy
-        </h3>
-        {olsPool.isActive ? (
-          <span className="inline-flex items-center rounded-full bg-emerald-900/60 px-2.5 py-0.5 text-xs font-medium text-emerald-300 ring-1 ring-emerald-700/50">
-            Active
-          </span>
-        ) : (
-          <span className="inline-flex items-center rounded-full bg-red-900/60 px-2.5 py-0.5 text-xs font-medium text-red-300 ring-1 ring-red-700/50">
-            Removed
-          </span>
-        )}
-      </div>
-
-      <dl className="text-sm space-y-4">
-        <div>
-          <h4 className="text-xs font-medium text-slate-500 uppercase tracking-wider mb-2">
-            Configuration
-          </h4>
-          <div className="grid grid-cols-2 sm:grid-cols-3 gap-4">
-            <Stat
-              label="Debt Token"
-              value={
-                !olsPool.debtToken
-                  ? "Unknown"
-                  : `${debtTokenSym} (${debtTokenSide})`
-              }
-            />
-            <Stat
-              label="Cooldown"
-              value={
-                cooldown > 0
-                  ? `${Math.floor(cooldown / 3600)}h ${Math.floor((cooldown % 3600) / 60)}m`
-                  : "None"
-              }
-            />
-            <Stat
-              label="OLS Contract"
-              value={<AddressLink address={olsPool.olsAddress} />}
-            />
-          </div>
-        </div>
-
-        <div className="border-t border-slate-800 pt-4">
-          <h4 className="text-xs font-medium text-slate-500 uppercase tracking-wider mb-2">
-            Activity
-          </h4>
-          <div className="grid grid-cols-2 sm:grid-cols-3 gap-4">
-            <Stat
-              label="OLS Rebalances"
-              value={String(olsPool.olsRebalanceCount)}
-            />
-            <Stat
-              label="Last Rebalance"
-              value={
-                lastRebalance > 0
-                  ? relativeTime(String(lastRebalance))
-                  : "Never"
-              }
-              title={
-                lastRebalance > 0
-                  ? formatTimestamp(String(lastRebalance))
-                  : undefined
-              }
-            />
-            <Stat
-              label="Protocol Fee Recipient"
-              value={
-                olsPool.protocolFeeRecipient ? (
-                  <AddressLink address={olsPool.protocolFeeRecipient} />
-                ) : (
-                  "Unknown"
-                )
-              }
-            />
-            <div className="col-span-2 sm:col-span-3">
-              <dt className="text-slate-400 mb-1">Cooldown Status</dt>
-              <dd className="text-white">
-                {lastRebalance === 0 ? (
-                  <span className="text-slate-500">Never rebalanced</span>
-                ) : cooldownReady ? (
-                  <span className="text-emerald-400">Ready to rebalance</span>
-                ) : (
-                  <div className="space-y-1">
-                    <span>
-                      {cooldownH}h {cooldownM}m remaining
-                    </span>
-                    <div className="w-full max-w-xs h-2 bg-slate-700 rounded-full overflow-hidden">
-                      <div
-                        className="h-full bg-indigo-500 rounded-full transition-all"
-                        style={{ width: `${cooldownPct}%` }}
-                      />
-                    </div>
-                  </div>
-                )}
-              </dd>
-            </div>
-          </div>
-        </div>
-
-        <div className="border-t border-slate-800 pt-4">
-          <h4 className="text-xs font-medium text-slate-500 uppercase tracking-wider mb-2">
-            Incentive Structure
-          </h4>
-          <div className="grid grid-cols-2 sm:grid-cols-4 gap-4">
-            <Stat
-              label="Expansion (Source)"
-              value={toPercent(olsPool.liquiditySourceIncentiveExpansion)}
-            />
-            <Stat
-              label="Contraction (Source)"
-              value={toPercent(olsPool.liquiditySourceIncentiveContraction)}
-            />
-            <Stat
-              label="Expansion (Protocol)"
-              value={toPercent(olsPool.protocolIncentiveExpansion)}
-            />
-            <Stat
-              label="Contraction (Protocol)"
-              value={toPercent(olsPool.protocolIncentiveContraction)}
-            />
-          </div>
-        </div>
-      </dl>
-    </div>
-  );
-}
-
-/**
- * Fetches OLS liquidity events scoped to the active OLS contract address,
- * preventing event mixing when a pool has been re-registered to a new OLS contract.
- */
-function OlsLiquidityEvents({
-  poolId,
-  olsAddress,
-  limit,
-  pool,
-  network,
-  search,
-  onSearchChange,
-}: {
-  poolId: string;
-  olsAddress: string | null;
-  limit: number;
-  pool: Pool | null;
-  network: ReturnType<typeof useNetwork>["network"];
-  search: string;
-  onSearchChange: (value: string) => void;
-}) {
-  const { getName, getTags } = useAddressLabels();
-  const searchQuery = normalizeSearch(search);
-  const [rawPage, setRawPage] = React.useState(1);
-
-  const handleSearchChange = React.useCallback(
-    (value: string) => {
-      onSearchChange(value);
-      setRawPage(1);
-    },
-    [onSearchChange],
-  );
-
-  const countQuery = olsAddress ? OLS_LIQUIDITY_EVENTS_COUNT : null;
-  const { data: countData, error: countError } = useGQL<{
-    OlsLiquidityEvent: { id: string }[];
-  }>(
-    countQuery,
-    olsAddress ? { poolId, olsAddress, limit: ENVIO_MAX_ROWS, offset: 0 } : {},
-  );
-  const lastKnownTotalRef = React.useRef(0);
-  const rawTotal = countData?.OlsLiquidityEvent?.length ?? 0;
-  if (rawTotal > 0) lastKnownTotalRef.current = rawTotal;
-  const total = countError ? lastKnownTotalRef.current : rawTotal;
-  const countCapped = rawTotal >= ENVIO_MAX_ROWS;
-
-  const totalPages = total > 0 ? Math.ceil(total / limit) : 1;
-  const page = Math.max(1, Math.min(rawPage, totalPages));
-  const isSearching = searchQuery.length > 0;
-  const searchFetchLimit =
-    total > 0 ? Math.min(total, SEARCH_MAX_LIMIT) : SEARCH_BOOTSTRAP_LIMIT;
-  const fetchLimit = isSearching ? searchFetchLimit : limit;
-  const fetchOffset = isSearching ? 0 : (page - 1) * limit;
-  const orderBy = useMemo(() => buildOrderBy("blockTimestamp", "desc"), []);
-
-  const gqlQuery = olsAddress ? OLS_LIQUIDITY_EVENTS_PAGE : null;
-  const { data, error, isLoading } = useGQL<{
-    OlsLiquidityEvent: OlsLiquidityEvent[];
-  }>(
-    gqlQuery,
-    olsAddress
-      ? { poolId, olsAddress, limit: fetchLimit, offset: fetchOffset, orderBy }
-      : {},
-  );
-
-  const events = data?.OlsLiquidityEvent ?? [];
-
-  const filteredEvents = useMemo(() => {
-    if (!searchQuery) return events;
-    return events.filter((e) => {
-      const givenSym = tokenSymbol(network, e.tokenGivenToPool);
-      const takenSym = tokenSymbol(network, e.tokenTakenFromPool);
-      const givenDec =
-        pool?.token0?.toLowerCase() === e.tokenGivenToPool.toLowerCase()
-          ? (pool?.token0Decimals ?? 18)
-          : (pool?.token1Decimals ?? 18);
-      const takenDec =
-        pool?.token0?.toLowerCase() === e.tokenTakenFromPool.toLowerCase()
-          ? (pool?.token0Decimals ?? 18)
-          : (pool?.token1Decimals ?? 18);
-      return matchesRowSearch(searchQuery, [
-        e.txHash,
-        e.direction === 0 ? "expand" : "contract",
-        ...addressSearchTerms(e.caller, getName, getTags),
-        formatWei(e.amountGivenToPool, givenDec),
-        givenSym,
-        formatWei(e.amountTakenFromPool, takenDec),
-        takenSym,
-      ]);
-    });
-  }, [events, searchQuery, pool, network, getName, getTags]);
-
-  return (
-    <>
-      {events.length > 0 && (
-        <TableSearch
-          value={search}
-          onChange={handleSearchChange}
-          placeholder="Search OLS events by tx, caller, direction, amount, or token..."
-          ariaLabel="Search OLS events"
-        />
-      )}
-      {error ? (
-        <OlsLiquidityTable
-          events={[]}
-          pool={pool}
-          network={network}
-          isLoading={false}
-          error={error}
-        />
-      ) : searchQuery && events.length > 0 && filteredEvents.length === 0 ? (
-        <EmptyBox message="No OLS events match your search." />
-      ) : (
-        <OlsLiquidityTable
-          events={filteredEvents}
-          pool={pool}
-          network={network}
-          isLoading={isLoading}
-          error={null}
-        />
-      )}
-      {!error && !isSearching && (
-        <Pagination
-          page={page}
-          pageSize={limit}
-          total={total}
-          onPageChange={setRawPage}
-        />
-      )}
-      {!error && countCapped && !isSearching && (
-        <p className="px-1 pt-1 text-xs text-amber-400">
-          Showing first {ENVIO_MAX_ROWS.toLocaleString()} OLS events — older
-          entries may exist beyond this page range.
-        </p>
-      )}
-      {!error && countCapped && isSearching && (
-        <p className="px-1 pt-1 text-xs text-amber-400">
-          Search covers the most recent {ENVIO_MAX_ROWS.toLocaleString()} OLS
-          events only.
-        </p>
-      )}
-      {!error && countError && !isSearching && (
-        <p className="px-1 pt-1 text-xs text-amber-400">
-          Could not load total count — pagination may be incomplete.
-        </p>
-      )}
-      {!error && countError && isSearching && (
-        <p className="px-1 pt-1 text-xs text-amber-400">
-          Could not load total count — search covers the most recent{" "}
-          {SEARCH_BOOTSTRAP_LIMIT.toLocaleString()} entries only.
-        </p>
-      )}
-    </>
-  );
-}
-
-export function OlsLiquidityTable({
-  events,
-  pool,
-  network,
-  isLoading,
-  error,
-}: {
-  events: OlsLiquidityEvent[];
-  pool: Pool | null;
-  network: ReturnType<typeof useNetwork>["network"];
-  isLoading: boolean;
-  error: Error | null;
-}) {
-  if (error) return <ErrorBox message={error.message} />;
-  if (isLoading) return <Skeleton rows={5} />;
-  if (events.length === 0)
-    return <EmptyBox message="No OLS liquidity events for this pool." />;
-
-  return (
-    <Table>
-      <thead>
-        <tr className="border-b border-slate-800 bg-slate-900/50">
-          <Th>Time</Th>
-          <Th>Direction</Th>
-          <Th align="right">Given to Pool</Th>
-          <Th align="right">Taken from Pool</Th>
-          <Th>Caller</Th>
-          <Th>Tx</Th>
-        </tr>
-      </thead>
-      <tbody>
-        {events.map((e) => {
-          const givenSym = tokenSymbol(network, e.tokenGivenToPool);
-          const takenSym = tokenSymbol(network, e.tokenTakenFromPool);
-          const givenDec =
-            pool?.token0?.toLowerCase() === e.tokenGivenToPool.toLowerCase()
-              ? (pool?.token0Decimals ?? 18)
-              : (pool?.token1Decimals ?? 18);
-          const takenDec =
-            pool?.token0?.toLowerCase() === e.tokenTakenFromPool.toLowerCase()
-              ? (pool?.token0Decimals ?? 18)
-              : (pool?.token1Decimals ?? 18);
-          return (
-            <Row key={e.id}>
-              <Td small muted title={formatTimestamp(e.blockTimestamp)}>
-                {relativeTime(e.blockTimestamp)}
-              </Td>
-              <td className="px-4 py-2">
-                {e.direction === 0 ? (
-                  <span className="inline-flex items-center rounded-full bg-emerald-900/60 px-2 py-0.5 text-xs font-medium text-emerald-300 ring-1 ring-emerald-700/50">
-                    EXPAND
-                  </span>
-                ) : (
-                  <span className="inline-flex items-center rounded-full bg-red-900/60 px-2 py-0.5 text-xs font-medium text-red-300 ring-1 ring-red-700/50">
-                    CONTRACT
-                  </span>
-                )}
-              </td>
-              <Td mono small align="right">
-                {formatWei(e.amountGivenToPool, givenDec)} {givenSym}
-              </Td>
-              <Td mono small align="right">
-                {formatWei(e.amountTakenFromPool, takenDec)} {takenSym}
-              </Td>
-              <SenderCell address={e.caller} />
-              <TxHashCell txHash={e.txHash} />
-            </Row>
-          );
-        })}
-      </tbody>
-    </Table>
   );
 }
